décomposition KmHelp + css
